ceil/floor/truncate/round return integers.
This CL supercedes https://codereview.chromium.org/11748016/

Review URL: https://codereview.chromium.org//12317107

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge@19810 260f80e4-7a28-3924-810f-c04153c831b5

+++ dart/samples/logo/logo.dart
@@ -43,7 +43,7 @@ class Color {
       s = d/(2 - 2*l);
     }
 
-    return new Color((h.round() % 360).toInt(), s, l);
+    return new Color(h.round() % 360, s, l);
   }
 
   factory Color.fromHex(String hex) => new Color.rgb(
